+++ lib/blocs/bloc.dart
@@ -6,6 +6,6 @@ export 'building_info_bloc.dart';
 export 'building_info_event.dart';
 export 'building_info_state.dart';
 
-export 'directions_ui_bloc.dart';
-export 'directions_ui_event.dart';
-export 'directions_ui_state.dart';
+export 'directions_bloc.dart';
+export 'directions_event.dart';
+export 'directions_state.dart';
